перенес процедуру инициализации бд, добавил стейт машину в FB бота
